Move information_schema collectors to collector package.
* collect.info_schema.innodb_metrics
* collect.info_schema.processlist
* collect.info_schema.query_response_time
* collect.info_schema.userstats
* collect.info_schema.tables
* collect.info_schema.tablestats
