+++ collector/info_schema.go
@@ -1,0 +1,4 @@
+package collector
+
+// Subsystem.
+const informationSchema = "info_schema"
